# shared/openapi/api-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CogniBoiler API Gateway",
    "description": "REST and WebSocket gateway of the CogniBoiler digital twin: sign-in and sessions, plant state, PLC commands, simulation control, alarms, history, audit and user administration. Errors are Problem Details (RFC 9457).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/commands/valve": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Send Valve Command",
        "description": "Send a manual valve command to the PLC; the PLC switches to MANUAL.",
        "operationId": "send_valve_command_api_v1_commands_valve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValveCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/commands/load": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Set Load Demand",
        "description": "Set the electrical load the PLC drives the unit to in AUTO.",
        "operationId": "set_load_demand_api_v1_commands_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadDemandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/commands/mode": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Set Control Mode",
        "description": "Switch the PLC between AUTO and MANUAL, or trip the unit.",
        "operationId": "set_control_mode_api_v1_commands_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/commands/setpoint": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Update Setpoints",
        "description": "Update live PLC setpoints and return the acceptance status.",
        "operationId": "update_setpoints_api_v1_commands_setpoint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/commands/reset": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Reset Emergency Stop",
        "description": "Reset the PLC emergency stop latch once its cause is gone; back to AUTO.\n\nThe reset is recorded under the authenticated user. An operator_id in the body no\nlonger overrides that; it is kept in the audit detail as a note.",
        "operationId": "reset_emergency_stop_api_v1_commands_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ResetRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandAckResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "description": "True if the command was accepted."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Rejection reason. Empty string when accepted=True.",
            "default": ""
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms",
            "description": "UTC epoch milliseconds."
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "timestamp_ms"
        ],
        "title": "CommandAckResponse",
        "description": "Response body for all command endpoints.\nMirrors CommandAck from the protobuf schema.\nIf accepted=False, reason explains why the command was rejected\n(e.g. value out of safe range, safety interlock active)."
      },
...
      "ControlModeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "auto",
              "manual",
              "estop"
            ],
            "title": "Mode",
            "description": "PLC operating mode to switch to."
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ControlModeRequest",
        "description": "Request body for POST /api/v1/commands/mode.\nRequires minimum role: operator. \"estop\" is a manual trip; only a reset clears it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoadDemandRequest": {
        "properties": {
          "load_w": {
            "type": "number",
            "maximum": 300000000.0,
            "minimum": 0.0,
            "title": "Load W",
            "description": "Electrical load target [W], 0 – 300 MW.",
            "examples": [
              300000000.0
            ]
          }
        },
        "type": "object",
        "required": [
          "load_w"
        ],
        "title": "LoadDemandRequest",
        "description": "Request body for POST /api/v1/commands/load.\nRequires minimum role: operator. The unit ramps to the new load at 30 MW/min."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "ProblemDetails",
        "description": "An error body (RFC 9457); members beyond these depend on the code."
      },
...
      "ResetRequest": {
        "properties": {
          "operator_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Operator Id",
            "description": "Operator identifier recorded in the reset audit trail.",
            "default": ""
          }
        },
        "type": "object",
        "title": "ResetRequest",
        "description": "Request body for POST /api/v1/commands/reset."
      },
...
      "SetpointRequest": {
        "properties": {
          "pressure_pa": {
            "type": "number",
            "maximum": 18500000.0,
            "minimum": 5000000.0,
            "title": "Pressure Pa",
            "description": "Target drum pressure [Pa]. Safe range: 50–185 bar.",
            "examples": [
              14000000.0
            ]
          },
          "water_level_m": {
            "type": "number",
            "maximum": 9.0,
            "minimum": 0.5,
            "title": "Water Level M",
            "description": "Target water level in drum [m].",
            "examples": [
              4.8
            ]
          },
          "steam_temp_k": {
            "type": "number",
            "maximum": 848.0,
            "minimum": 400.0,
            "title": "Steam Temp K",
            "description": "Target steam temperature [K]. Safe range: 400–848 K.",
            "examples": [
              811.0
            ]
          }
        },
        "type": "object",
        "required": [
          "pressure_pa",
          "water_level_m",
          "steam_temp_k"
        ],
        "title": "SetpointRequest",
        "description": "Request body for POST /api/v1/commands/setpoint.\nRequires minimum role: engineer.\nAll three setpoints must be provided together — the cascade PID\ncontroller uses them as a consistent set."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValveCommandRequest": {
        "properties": {
          "fuel_valve": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fuel Valve",
            "description": "Fuel valve position [0.0 = closed, 1.0 = fully open].",
            "examples": [
              0.75
            ]
          },
          "feedwater_valve": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Feedwater Valve",
            "description": "Feedwater valve position [0.0 – 1.0].",
            "examples": [
              0.6
            ]
          },
          "steam_valve": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Steam Valve",
            "description": "Turbine admission valve position [0.0 – 1.0].",
            "examples": [
              0.5
            ]
          },
          "spray_valve": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Spray Valve",
            "description": "Attemperator spray valve position [0.0 – 1.0]; omitted keeps it.",
            "examples": [
              0.25
            ]
          }
        },
        "type": "object",
        "required": [
          "fuel_valve",
          "feedwater_valve",
          "steam_valve"
        ],
        "title": "ValveCommandRequest",
        "description": "Request body for POST /api/v1/commands/valve.\nRequires minimum role: operator.\nAll three valve positions must be provided together to avoid\npartial state updates that could destabilise the control loop."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}